Subscription-based presence (#363)
* Eliminate setPresence method

Instead you register your presence with a subscription. When you change rooms, you subscribe for the new room, then unsubscribe from the old one; when you close the tab or go to a non-chat view, you unsubscribe as well. In both cases we update the database to show you as not being in the room on that connection. This lets us generate the part message immediately when you leave from all clients, without spuriously generating it if you leave from one client but are still present in another.
For the time being this means we're not generating the foreground signal, so you always show up as being in the background.

* Track presence in jitsi chat

Doesn't work if you pop the chat out, but otherwise tracks presence.
Added client side transformation of new presence table to turn O(N*M) into O(n+m) on puzzles and sessions during page render, since having chat and jitsi presence was going to increase m.
Foreground now means present in Jitsi even if not in chat. Background is in chat and not jitsi. Made background somewhat transparent, put foreground first, and separated thim if both are present.
Fixes #315

* Remove extra client-side logging

* Don't store subscription info on connection

It's unnecessary since we're sure the onStop will get called.

* Defer presence removal by a few seconds

If you switch from side-by-side to full screen chat it shouldn't show you parting and joining.

* Fix /list command