Private message drawer (#384)
* First pass at making private messages in all channels globally available

Adds a panel to the avatar dropdown (which was otherwise only useful for logging out) that contains all the private messages to you ever.
Adds a virtual room for LastRead purposes called "private" which is when you last marked private messages as read. Any private messages since you did that are counted in a red badge over your avatar.

* Change background color of unread messages

* Private messages are also read if their own room LastRead is past them

Instead of just the synthetic 'private' LastRead entry.

* Fix last read app test

Technically just fixes the latency compensation.