Use @ to mark mentions and track them as first class message fields (#400)
* Basic implementation of mentions

Doesn't provide an interface to set the mention property on a new message, but if present it will be rendered appropriately.

* @ plus a nickname mentions the nickname

adding it to the mention list of the message. The nickname has to match an existing user id.

* Make waitForMethods wait for methods

* Don't highlight text when mention is present

Use the mention field instead for mention count.

* Highlight actions that mention you

* Make bot @mention in replies

* Subscribe to messages that mention you

Adds index to Messages collection to support the query

* Add mentions tab alongside private messages tab

Mentions are available anywhere with link to room of origin. If no private messages and any mentions, an orange badge shows the number of mentions up to 9. The mark read button applies to both.

* Notify on mentions

if private-messages stream is enabled

* Refactor template for location of mention/private message

* Add back pop-out link to message/mention room