Logistics page (#709)
* First pass at logistics page

Shows metas as blocks. Uncategorized not shown yet. Can't edit anything from the page yet.

* Add callin queue to logistics page

* Set title on logistics page

* Logistics page puzzle grid should scroll

* Transition height instead of max-height

Also have an overshoot and settle.

* Settle smoothly after overshoot

* Make logistics page part of expert mode

* Fix tests

* Switching to logistics doesn't cause a reload

* Both callins and logistics in header and sidebar

* Apply stuck style to puzzles

Was applying it to everything if the meta was stuck

* Set link correctly on callins

* Make puzzle and meta names links

* Clear top right when navigating to logistics

* Size meta names properly

* Increase minimum column width

* Add top right panel to logistics view

* Make answer type selector drop up on logistics page

* Remove callin control from table header on logistics

* Make solved metas transparent

* Smaller margin below starred messages on logistics

* Allow modifying starred messages on logistics

* Add chat/jitsi presence count to puzzles in logistics

* Don't use percent for opacity.

* Remove underscore from logistics puzzle presence

Everything in that file is private to the page, so prefixing anything is pointless.

* Remove logging

* Keep top right panel around when reloading

It was trying to load it immediately after a code refresh, but since the template is in a dynamic bundle it wasn't loaded yet. Templates loading isn't reactive, and setting the session key to the same value doesn't invalidate dependencies.

* Add non-feeder, non-meta puzzles to logistics

* Delete logistics from components

It's in pages.

* Put onduty control in a btn-toolbar.

* Fix link text specificity on dark mode

* Add logos for puzzles with calendar events

* Add event icon to metas

* Fix overflow

* Pair standalone puzzles

Puzzles within a round that don't feed a meta will be grouped in pairs, like puzzles within a meta, so that we don't waste half a column of space.

* Allow creating things from logistics

Doesn't allow moving standalones into metas or creating puzzles directly in metas yet.

* Allow making new puzzles directly in metas

* Fix adding first standalone to round

* Remove logging

* Add tooltip to new puzzle in meta button

* Drag and drop link to make puzzle/round

Dropping a link onto the New Round button makes a round. Dropping a link onto a round name under the New Meta button makes a new meta in that round. Dropping a link onto a round name under the New Puzzle button make a standalone puzzle in that round. Dropping a link on an existing meta adds a new feeder to that meta. In all cases, the link to the new puzzle is the href of the link being dragged, and the name is the text of the link.

* Add highlight when dragging over legal drop target

* Allow deleting puzzles by dragging to trash

* Don't use ReactiveDict for dragged puzzle

It will only change on events, and something else will always change too

* Blur out puzzle when dragging it to trash

* Allow feeding puzzle into meta with drag and drop

* Add tooltip to delete target

Since it's "disabled", the explanation that you drop stuff on it might be helpful.

* Allow unfeeding meta by dragging it to empty space

* Fix --num-feeders for standalones

* Add edit dialog to logistics

Edit button is visible when hovering over the puzzle/meta.

* Add logistics view to oncall doc

* Fix links to dynamic settings

* Show some of the header buttons in puzzle info from logistics

* Add escape to close puzzle edit modal

Not clear why it wasn't working

* Add calendar link to logistics top right panel

* Show calendar events on logistics

If not attached, dragging them to a puzzle or meta attaches them.

* Move the calendar events to the right side of logistics

* Update oncall doc with calendar instructions

* Move note about callback number to onduty section

* Eliminate callins page

Also fixes some tests that were apparently order-dependent when moving the callin queue tests to the logistics directory.

* Move callins_table to logistics template

Since the callins page is gone, we don't need it to take up space in the base bundle.

* Allow fixing drive from other pages

Both logistics and puzzle info

* Don't make logistics 100% height

Because we set both top and bottom, its height is as big as possible, and setting height to 100% means the bottom of the call-in queue can be scrolled off.

* Fix occasional weird bottom border

Seems like it was caused by standalone groups being slightly smaller than they should have been.

* Redirect from callins to logistics

In case someone doesn't read the updates.

* Missing trailing newlines